+++ test/argus/dl_declarations_test.exs
@@ -144,7 +144,8 @@ defmodule Argus.DlDeclarationsTest do
            ignored_error_result implements_behaviour monitor_call trap_exit),
       ets: ~w(dynamic_child ets_new ets_op ets_option function_def implements_behaviour
               supervisor_child),
-      gen_statem: ~w(statem_initial statem_module statem_state statem_transition),
+      gen_statem:
+        ~w(function_def statem_event_catchall statem_event_clause statem_info_catchall statem_initial statem_module statem_state statem_timeout statem_transition),
       message_contract:
         ~w(async_cast callback_tag callback_total def_use function_def implements_behaviour
            literal_value remote_call sync_call tuple_literal),